FIX: 64-bit - Fixed places where pointers where expected to fit into 32 bit integer - Fixed some compiler warnings of converting 64 bit to 32 bit (loosing precision) - Fixed pointer arithmetic in x_memory_std - Fixed TARGET_32BIT and TARGET_64BIT - Fixed xsize_t and X_PTR_SIZED_INT, in 32 bit mode this is a 32 bit integer,         in 64 bit mode this is a 64 bit integer. - Fixed array size of static array in singleton static instantiation

+++ source/main/cpp/x_qsort.cpp
@@ -139,15 +139,15 @@ namespace xcore
 		}
 
 		pn = (xbyte*) a + n * es;
-		r = __xqsort_MIN(pa - (xbyte*)a, pb - pa);
+		r = (s32)(__xqsort_MIN(pa - (xbyte*)a, pb - pa));
 		__xqsort_VecSwap((xbyte*)a, pb - r, r);
-		r = __xqsort_MIN(pd - pc, pn - pd - es);
+		r = (s32)(__xqsort_MIN(pd - pc, pn - pd - es));
 		__xqsort_VecSwap(pb, pn - r, r);
 
-		if ((r = pb - pa) > es)
+		if ((r = (s32)(pb - pa)) > es)
 			xqsort((xbyte*)a, r / es, es, cmp, data);
 	
-		if ((r = pd - pc) > es)
+		if ((r = (s32)(pd - pc)) > es)
 		{ 
 			// Iterate rather than recurse to save stack space
 			a = pn - r;
